requirements.txt
asgiref==3.2.5
certifi==2020.6.20
chardet==3.0.4
Django==3.0.4
django-cors-headers==3.5.0
djangorestframework==3.12.1
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
jdcal==1.4.1
openpyxl==3.0.3
pytz==2019.3
requests==2.24.0
selenium==3.141.0
sqlparse==0.3.1
stripe==2.54.0
urllib3==1.25.10
whitenoise==5.2.0
